pyproject.toml
[tool.poetry]
name = "cnn_with_cifar10"
version = "0.1.0"
description = ""
authors = ["r-matsuzaka <matsuzakaryo@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8.6"
Pillow = "^8.4.0"
isort = "^5.10.1"
mypy = "^0.910"
pytest = "^6.2.5"
black = "^21.11b1"
pylint = "^2.11.1"
pytest-cov = "^3.0.0"
tensorflow = "^2.7.0"
matplotlib = "^3.5.0"
natsort = "^8.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
